Extra (#4)
* ADD Boss

* Fix bugs

* MOD folder structure

* ADD events to boss

* Add Hero

* Fix ending

* Upgrade texts component

+++ SpaceHell/PlayerController.h
@@ -15,7 +15,7 @@ constexpr glm::vec2 PLAYER_SHIELD_SIZE(96.0f, 96.0f);
 constexpr glm::vec2 PLAYER_SHIELD_OFFSET(-16.0f, -10.0f);
 constexpr glm::vec2 PLAYER_COllIDER_OFFSET(0.0f, 5.0f);
 constexpr int PLAYER_SHIELD_NUMBER = 12;
-constexpr int PLAYER_BULLET_SPEED = 500;
+constexpr float PLAYER_BULLET_SPEED = 500.0f;
 constexpr glm::vec2 PLAYER_BULLET_SIZE(12.0f, 12.0f);
 constexpr float PLAYER_BULLET_RELOAD = 1.0f;
 constexpr int PLAYER_BULLET_DEFAULT_NUMBER = 1;
